backend/google_vm_recognition/requirements.txt
Flask==1.0.2
opencv-python
numpy
keras_applications
tensorflow
keras_vggface
mtcnn
scipy
pandas
google-cloud-storage
google-cloud-firestore
functions-framework
h5py < 3.0.0